Move KeyCodeExtensions into Extensions namespace
