+++ Assets/Scripts/Input/KeyboardTarget.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using PAC.DataStructures;
+using PAC.Extensions;
 using PAC.KeyboardShortcuts;
 using UnityEngine;
 using UnityEngine.Events;
